[Feat] - 데이터 타입에 의한 버그 수정

+++ core/network/src/main/java/com/threedollar/network/data/store/Feedback.kt
@@ -8,5 +8,5 @@ data class Feedback(
     @SerializedName("feedbackType")
     val feedbackType: String? = "",
     @SerializedName("ratio")
-    val ratio: Int? = 0
+    val ratio: Double? = 0.0
 )
